refactor: Replace FormFieldEditor with cleaner FieldCard implementation
- Extract FieldCard component from InteractiveFormPanel to molecules package
- Replace FormFieldEditor usage in FormFieldList with FieldCard
- Add comprehensive Storybook stories for FieldCard component
- Remove deprecated FormFieldEditor files and stories
- Improve UX with inline editing, keyboard navigation, and better validation
- Maintain all existing functionality while providing cleaner architecture
